requirements.txt
connexion[swagger-ui] == 2.13.0
python_dateutil == 2.8.2
setuptools >= 21.0.0
weasyprint == 54.3
python-magic == 0.4.25
